CSLU2024 - UM/Shen/challenges/blindsql/app/requirements.txt
Flask==3.0.0
Waitress==2.1.2
mysql-connector-python